+++ gov/utils.go
@@ -22,9 +22,9 @@ func GetProposalIDFromInput(bin *utils.Binary, args []string) (int, error) {
 			return 0, errors.Wrap(err, "error querying latest proposal ID")
 		}
 	case 1:
-		proposalID, err = strconv.Atoi(args[2])
+		proposalID, err = strconv.Atoi(args[0])
 		if err != nil {
-			return 0, errors.Wrapf(err, "error converting proposal ID %s to integer", args[2])
+			return 0, errors.Wrapf(err, "error converting proposal ID %s to integer", args[0])
 		}
 	default:
 		return 0, fmt.Errorf("invalid number of arguments; expected 0 or 1; got %d", len(args))